Work around Postgres returning all columns instead of just generated keys on INSERT
Postgres' JDBC driver isn't aware of what columns were automatically
generated in an INSERT, so it returns all the rows. This was causing
unmapped columns to be parsed in all result sets, and if one of them was
of an unknown type, it would cause an exception.

Fix the problem by having callers specify what keys they want from the
insert, and only parsing and returning the interesting ones, all the way
from the mapping calls downwards to the JDBC insertion code.